+++ Libplanet.Benchmarks/MineBlock.cs
@@ -61,7 +61,7 @@ namespace Libplanet.Benchmarks
         public void MakeOneTransactionWithActions()
         {
             var privateKey = new PrivateKey();
-            var address = privateKey.PublicKey.ToAddress();
+            var address = privateKey.ToAddress();
             var actions = new[]
             {
                 new DumbAction(address, "foo"),
@@ -84,7 +84,7 @@ namespace Libplanet.Benchmarks
             for (var i = 0; i < 10; i++)
             {
                 var privateKey = new PrivateKey();
-                var address = privateKey.PublicKey.ToAddress();
+                var address = privateKey.ToAddress();
                 var actions = new[]
                 {
                     new DumbAction(address, "foo"),